% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_error_correction.R
\name{estimate_error_correction}
\alias{estimate_error_correction}
\title{Error correction estimator}
\arguments{
\item{y}{Character}

\item{X}{Character (vector)}

\item{data}{Data.table}
}
\value{
Returns the long-run estimates in a data.table format
}
\description{
Error correction estimator
}
