% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weasel_words.R
\name{find_weasels}
\alias{find_weasels}
\title{counts the number of 'weasel words' in a document.}
\usage{
find_weasels(txt, weasels, outfile = FALSE)
}
\arguments{
\item{txt}{text to evaluate.}

\item{weasels}{list of words or phrases to look for.}
}
\value{
a list of weasel phrases found and their frequency.
}
\description{
counts the number of 'weasel words' in a document.
}

