% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/researchTools.R
\docType{package}
\name{researchTools}
\alias{researchTools}
\alias{researchTools-package}
\title{researchTools: A package of various helper functions for research and writing.}
\description{
The researchTools package provides \code{get_paper} and \code{find_weasels}.
}
\section{writing functions}{

The writing sections include \code{get_paper} and \code{find_weasels}.
\code{get_paper}, given a DOI, adds a bibtex citation to a bib file.
\code{find_weasels} evaluates text and compiles a count of words that authors
may want to reduce or remove.
}

