% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{weasel_words}
\alias{weasel_words}
\title{Words to avoid in writing.}
\format{A character array of words.}
\usage{
weasel_words
}
\description{
A list of words and phrases that should be avoided in writing. Named after
Matt Might's \link{weasel words script}{http://matt.might.net/articles/shell-scripts-for-passive-voice-weasel-words-duplicates/}.
Also from Dierdre McCloskey's list of words to avoid in Economical Writing.
}
\keyword{datasets}

